feature: role based access added to company

+++ src/api-gateway/src/company/company.controller.ts
@@ -59,6 +59,7 @@ import { JwtService } from '@nestjs/jwt';
 import { BooleanResponseDto } from '../shared/dtos/api-response.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { JobStatusAllResponseDto } from '../job/dto/job-responses.dto';
+import { EmployeeService } from 'src/employee/employee.service';
 
 const className = 'Company';
 
@@ -68,6 +69,7 @@ export class CompanyController {
   constructor(
     private readonly companyService: CompanyService,
     private readonly jwtService: JwtService,
+    private readonly employeeService: EmployeeService,
   ) {}
   validateObjectId(id: string | Types.ObjectId): boolean {
     if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -264,26 +266,37 @@ export class CompanyController {
   async update(
     @Headers() headers: any,
     @Param('cid') cid: string,
-    @Body() updateCompanyDto: UpdateCompanyDto,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyDto: UpdateCompanyDto;
+    },
   ) {
-    try {
-      validateObjectId(cid);
-      const userId = extractUserId(this.jwtService, headers);
+    const currentEmployee = await this.employeeService.findById(
+      body.currentEmployeeId,
+    );
+    if (currentEmployee.role.permissionSuite.includes('view all employees')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
 
-      const companyId = new Types.ObjectId(cid);
-      const updatedCompany = await this.companyService.update(
-        userId,
-        companyId,
-        updateCompanyDto,
-      );
-      return {
-        data: updatedCompany,
-      };
-    } catch (e) {
-      throw new HttpException(
-        'internal server error',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+        const companyId = new Types.ObjectId(cid);
+        const updatedCompany = await this.companyService.update(
+          userId,
+          companyId,
+          body.updateCompanyDto,
+        );
+        return {
+          data: updatedCompany,
+        };
+      } catch (e) {
+        throw new HttpException(
+          'internal server error',
+          HttpStatus.INTERNAL_SERVER_ERROR,
+        );
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -304,22 +317,30 @@ export class CompanyController {
     @Headers() headers: any,
     @Param('cid') companyId: string,
     @UploadedFile() file: Express.Multer.File,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
   ) {
-    try {
-      validateObjectId(companyId);
-      const userId = extractUserId(this.jwtService, headers);
-      return {
-        data: await this.companyService.updateLogo(
-          userId,
-          new Types.ObjectId(companyId),
-          file,
-        ),
-      };
-    } catch (e) {
-      throw new HttpException(
-        'internal server error',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+    const currentEmployee = await this.employeeService.findById(
+      body.currentEmployeeId,
+    );
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(companyId);
+        const userId = extractUserId(this.jwtService, headers);
+        return {
+          data: await this.companyService.updateLogo(
+            userId,
+            new Types.ObjectId(companyId),
+            file,
+          ),
+        };
+      } catch (e) {
+        throw new HttpException(
+          'internal server error',
+          HttpStatus.INTERNAL_SERVER_ERROR,
+        );
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -340,18 +361,29 @@ export class CompanyController {
     description: `The _id attribute of the ${className}`,
   })
   @Delete(':cid')
-  async remove(@Headers() headers: any, @Param('cid') cid: string) {
-    try {
-      validateObjectId(cid);
-      const userId = extractUserId(this.jwtService, headers);
-      const objectId = new Types.ObjectId(cid);
-      await this.companyService.deleteCompany(userId, objectId);
-      return { data: true };
-    } catch (e) {
-      throw new HttpException(
-        'Internal Server Error',
-        HttpStatus.SERVICE_UNAVAILABLE,
-      );
+  async remove(
+    @Headers() headers: any,
+    @Param('cid') cid: string,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(
+      body.currentEmployeeId,
+    );
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
+        const objectId = new Types.ObjectId(cid);
+        await this.companyService.deleteCompany(userId, objectId);
+        return { data: true };
+      } catch (e) {
+        throw new HttpException(
+          'Internal Server Error',
+          HttpStatus.SERVICE_UNAVAILABLE,
+        );
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -373,17 +405,35 @@ export class CompanyController {
   @Delete('/emp')
   async removeEmployee(
     @Headers() headers: any,
-    @Body() deleteEmployeeDto: DeleteEmployeeFromCompanyDto,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      deleteEmployeeDto: DeleteEmployeeFromCompanyDto;
+    },
   ) {
-    try {
-      const userId = extractUserId(this.jwtService, headers);
-      await this.companyService.deleteEmployee(userId, deleteEmployeeDto);
-      return { data: true };
-    } catch (e) {
-      throw new HttpException(
-        'Internal Server Error',
-        HttpStatus.SERVICE_UNAVAILABLE,
-      );
+    const currentEmployee = await this.employeeService.findById(
+      body.currentEmployeeId,
+    );
+    if (currentEmployee.role.permissionSuite.includes('remove any employees')) {
+      try {
+        const userId = extractUserId(this.jwtService, headers);
+        await this.companyService.deleteEmployee(
+          userId,
+          body.deleteEmployeeDto,
+        );
+        return { data: true };
+      } catch (e) {
+        throw new HttpException(
+          'Internal Server Error',
+          HttpStatus.SERVICE_UNAVAILABLE,
+        );
+      }
+    } else if (
+      currentEmployee.role.permissionSuite.includes('remove employees under me')
+    ) {
+      //TODO
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -431,22 +481,33 @@ export class CompanyController {
   @Patch('statuses')
   async updateStatusOrder(
     @Headers() headers: any,
-    @Body() updateCompanyJobStatusesDto: UpdateCompanyJobStatusesDto,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyJobStatusesDto: UpdateCompanyJobStatusesDto;
+    },
   ) {
-    try {
-      const userId = extractUserId(this.jwtService, headers);
-      const statusArr = new UpdateCompanyJobStatuses(
-        updateCompanyJobStatusesDto,
-      );
-      return {
-        data: await this.companyService.updateCompanyStatuses(
-          userId,
-          updateCompanyJobStatusesDto.employeeId,
-          statusArr,
-        ),
-      };
-    } catch (e) {
-      throw e;
+    const currentEmployee = await this.employeeService.findById(
+      body.currentEmployeeId,
+    );
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        const userId = extractUserId(this.jwtService, headers);
+        const statusArr = new UpdateCompanyJobStatuses(
+          body.updateCompanyJobStatusesDto,
+        );
+        return {
+          data: await this.companyService.updateCompanyStatuses(
+            userId,
+            body.updateCompanyJobStatusesDto.employeeId,
+            statusArr,
+          ),
+        };
+      } catch (e) {
+        throw e;
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 }
